% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDE_chemical_data.R
\docType{data}
\name{GSDE_Soil_Chemical_Properties}
\alias{GSDE_Soil_Chemical_Properties}
\title{The Global Soil Dataset for Earth System Modeling - chemical properties}
\format{
A .tif raster file
}
\usage{
data(GSDE_Soil_Chemical_Properties)
}
\description{
A  raster file involving 25 soil chemical variables.Detail at: http://globalchange.bnu.edu.cn/research/soilwd.jsp
}
\examples{
\dontrun{
data(GSDE_Soil_Chemical_Properties)
require(raster)
plot(GSDE_Soil_Chemical_Properties)
}
}
\keyword{datasets}
