% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDE_physical_data.R
\docType{data}
\name{GSDE_Soil_Physical_Properties}
\alias{GSDE_Soil_Physical_Properties}
\title{The Global Soil Dataset for Earth System Modeling - physical properties}
\format{
A .tif raster file
}
\usage{
data(GSDE_Soil_Physical_Properties)
}
\description{
A  raster file involving 9 soil physical variables.Detail at: http://globalchange.bnu.edu.cn/research/soilwd.jsp
}
\examples{
\dontrun{
data(GSDE_Soil_Physical_Properties)
require(raster)
plot(GSDE_Soil_Physical_Properties)
}
}
\keyword{datasets}
